# Conflicts:
#	empedocle/empedocle/src/main/java/it/unifi/ing/stlab/empedocle/actions/health/examination/statistics/StatisticsDownload.java
#	empedocle/empedocle/src/main/java/it/unifi/ing/stlab/empedocle/actions/health/examination/statistics/excel/POIDataExporter.java
#	empedocle/empedocle/src/main/java/it/unifi/ing/stlab/empedocle/actions/health/examination/statistics/export/config/ExcelConfig.java
#	empedocle/empedocle/src/main/java/it/unifi/ing/stlab/empedocle/actions/health/examination/statistics/export/visitor/StatisticsPathGeneratorVisitor.java
#	empedocle/empedocle/src/main/webapp/pages/examination/examination-list.xhtml
#	empedocle/empedocle/target/classes/it/unifi/ing/stlab/empedocle/actions/factquery/FactQueryEdit.class
#	empedocle/empedocle/target/classes/it/unifi/ing/stlab/empedocle/actions/health/examination/ExaminationList.class
#	empedocle/empedocle/target/classes/it/unifi/ing/stlab/empedocle/actions/health/examination/types/ExaminationTypeEdit.class
#	empedocle/empedocle/target/classes/it/unifi/ing/stlab/empedocle/actions/patients/PatientEdit.class
#	empedocle/empedocle/target/classes/it/unifi/ing/stlab/empedocle/scripts/reconciliator/OfflinePatientReconciliator$MySQLConnector.class
#	empedocle/empedocle/target/classes/it/unifi/ing/stlab/empedocle/visitor/factquery/BuildExpressionVisitor.class
#	empedocle/empedocle/target/empedocle-1.0.0.war
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/classes/it/unifi/ing/stlab/empedocle/actions/factquery/FactQueryEdit.class
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/classes/it/unifi/ing/stlab/empedocle/actions/health/examination/ExaminationList.class
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/classes/it/unifi/ing/stlab/empedocle/actions/health/examination/types/ExaminationTypeEdit$2.class
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/classes/it/unifi/ing/stlab/empedocle/actions/health/examination/types/ExaminationTypeEdit$3.class
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/classes/it/unifi/ing/stlab/empedocle/actions/health/examination/types/ExaminationTypeEdit$4.class
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/classes/it/unifi/ing/stlab/empedocle/actions/health/examination/types/ExaminationTypeEdit.class
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/classes/it/unifi/ing/stlab/empedocle/actions/patients/PatientEdit.class
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/classes/it/unifi/ing/stlab/empedocle/scripts/reconciliator/OfflinePatientReconciliator$MySQLConnector.class
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/classes/it/unifi/ing/stlab/empedocle/visitor/factquery/BuildExpressionVisitor.class
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/empedocle-viewers-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-commons-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-entities-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-fact-query-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-filters-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-patients-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-reflection-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-reflection-crud-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-reflection-dsl-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-reflection-impl-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-users-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-view-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-view-controllers-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-view-crud-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/WEB-INF/lib/stlab-view-dsl-1.0.0.jar
#	empedocle/empedocle/target/empedocle-1.0.0/pages/examination/examination-list.xhtml
#	empedocle/empedocle/target/test-classes/it/unifi/ing/stlab/empedocle/model/health/coding/atc/ATCCodeJpaImporter.class
#	empedocle/empedocle/target/test-classes/it/unifi/ing/stlab/empedocle/model/health/coding/icd9cm/ICD9CMCodeJpaImporter.class
#	stlab-modules/stlab-entities/src/main/java/it/unifi/ing/stlab/entities/model/composite/CompositeLink.java
#	stlab-modules/stlab-fact-query/target/classes/it/unifi/ing/stlab/factquery/builder/FactQueryBuilder.class
#	stlab-modules/stlab-reflection-dsl/target/classes/it/unifi/ing/stlab/reflection/dsl/BaseTypeParser.class
#	stlab-modules/stlab-view-controllers/target/classes/it/unifi/ing/stlab/view/controllers/FileUploadBean.class
#	stlab-modules/stlab-view-dsl/target/classes/it/unifi/ing/stlab/view/dsl/BaseVistaParser.class

+++ empedocle/empedocle-viewers/src/main/java/it/unifi/ing/stlab/empedocle/view/controllers/geriatrics/MMSEController.java
@@ -1,12 +1,5 @@
 package it.unifi.ing.stlab.empedocle.view.controllers.geriatrics;
 
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
-
-import javax.enterprise.context.RequestScoped;
-import javax.inject.Named;
-
 import it.unifi.ing.stlab.entities.utils.ClassHelper;
 import it.unifi.ing.stlab.reflection.model.facts.Fact;
 import it.unifi.ing.stlab.reflection.model.facts.QualitativeFact;
@@ -16,6 +9,12 @@ import it.unifi.ing.stlab.view.controllers.ContainerController;
 import it.unifi.ing.stlab.view.model.Viewer;
 import it.unifi.ing.stlab.view.model.links.TypeSelector;
 import it.unifi.ing.stlab.view.model.links.ViewerLink;
+
+import javax.enterprise.context.RequestScoped;
+import javax.inject.Named;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.List;
 
 @Named
 @RequestScoped
@@ -86,7 +85,7 @@ public class MMSEController extends ContainerController {
 				int scoreValue = evaluateScore( facts );
 				double adjustedValue = scoreValue + getCorrectionFactor( 
 						ageIntervalFact.getPhenomenon().getName() + " " + educationLevelFact.getPhenomenon().getName() );
-				scoreFact.getQuantity().setValue( (double) adjustedValue );
+				scoreFact.getQuantity().setValue(adjustedValue);
 				
 				return String.valueOf( adjustedValue );
 			} else {
